Update README.md
Flappy Bird Python is a recreation of the classic game using Pygame library for graphics and Python for logic. Players navigate a bird through obstacles, tapping to keep it airborne. Built with Pygame for graphics rendering and Python for game logic, it's a fun project for Python enthusiasts to explore game development.